Refactoring: Payment 페이지 구조 리팩토링하기
- useState와 useEffect없이 Button렌더링하기
- payments로 사용하던 경로를 payment로 변경하기
- 하위 컴포넌트 이름 미리 작성해두기
