+++ src/pages/Payment/components/PayAmountInformation.tsx
