+++ src/pages/Payment/components/PayTermInformation.tsx
