+++ src/pages/Payment/components/index.ts
