+++ src/pages/Payment/index.style.ts
