+++ src/pages/index.ts
@@ -1,5 +1,5 @@
 export { default as Home } from './Home';
-export { default as Payments } from './Payments';
+export { default as Payment } from './Payment';
 export { default as CardList } from './CardList';
 export { default as CardRegist } from './CardRegist';
 export { default as CardAlias } from './CardAlias';
