+++ src/router/router.tsx
@@ -1,12 +1,13 @@
 import { createBrowserRouter } from 'react-router-dom';
-import { Home, Payments, CardList, CardRegist, CardAlias } from '../pages';
+
+import { Home, Payment, CardList, CardRegist, CardAlias } from '../pages';
 
 const router = createBrowserRouter([
   {
     element: <Home />,
     path: '/',
     children: [
-      { element: <Payments />, path: '/payments' },
+      { element: <Payment />, path: '/Payment' },
       { element: <CardList />, path: '/card-list' },
       { element: <CardRegist />, path: '/card-regist' },
       { element: <CardAlias />, path: '/card-alias/:id' },
